Move distribution sample generation into conftest and WIP performance assessment
